refactor: move files used by generator CLI to their own directory
