+++ packages/next-rest-framework/src/generate/constants.ts
