+++ packages/next-rest-framework/src/generate/generate.ts
