+++ packages/next-rest-framework/src/generate/utils.ts
