+++ packages/next-rest-framework/src/generate/validate.ts
